some changes to the reporter and curl_check
