--- tests/results-data.deic.dk-151023-114836
@@ -1,115 +1,0 @@
-{
-    "data.deic.dk": {
-        "fileTinkerDownload": [
-            {
-                "scenario": {
-                    "fileTinkerDownload_filesize": 300000000, 
-                    "fileTinkerDownload_tinker_wait": 1
-                }, 
-                "scenarioid": "0", 
-                "results": [
-                    {
-                        "worker": "worker0", 
-                        "sync_time": "0:00:15.061000"
-                    }, 
-                    {
-                        "worker": "worker1", 
-                        "sync_time": "0:00:12.993000"
-                    }, 
-                    {
-                        "errors": [
-                            " error_check(md5sum(fn) == md5) failed in tinkerer() [\"/home/piotr/DeIC/smashbox-reporter/lib/test_fileTinkerDownload.py\" at line 68]"
-                        ], 
-                        "worker": "tinkerer", 
-                        "sync_time": "0:00:00.000000"
-                    }
-                ], 
-                "timeid": "151023-114836", 
-                "loopid": "1", 
-                "total_exec_time": "0:00:32.801060"
-            }, 
-            {
-                "scenario": {
-                    "fileTinkerDownload_filesize": 300000000, 
-                    "fileTinkerDownload_tinker_wait": 2
-                }, 
-                "scenarioid": "1", 
-                "results": [
-                    {
-                        "worker": "worker0", 
-                        "sync_time": "0:00:14.231000"
-                    }, 
-                    {
-                        "worker": "worker1", 
-                        "sync_time": "0:00:12.029000"
-                    }, 
-                    {
-                        "errors": [
-                            " error_check(md5sum(fn) == md5) failed in tinkerer() [\"/home/piotr/DeIC/smashbox-reporter/lib/test_fileTinkerDownload.py\" at line 68]"
-                        ], 
-                        "worker": "tinkerer", 
-                        "sync_time": "0:00:00.000000"
-                    }
-                ], 
-                "timeid": "151023-114909", 
-                "loopid": "1", 
-                "total_exec_time": "0:00:29.776046"
-            }, 
-            {
-                "scenario": {
-                    "fileTinkerDownload_filesize": 300000000, 
-                    "fileTinkerDownload_tinker_wait": 3
-                }, 
-                "scenarioid": "2", 
-                "results": [
-                    {
-                        "worker": "worker0", 
-                        "sync_time": "0:00:14.608000"
-                    }, 
-                    {
-                        "worker": "worker1", 
-                        "sync_time": "0:00:11.765000"
-                    }, 
-                    {
-                        "errors": [
-                            " error_check(md5sum(fn) == md5) failed in tinkerer() [\"/home/piotr/DeIC/smashbox-reporter/lib/test_fileTinkerDownload.py\" at line 68]"
-                        ], 
-                        "worker": "tinkerer", 
-                        "sync_time": "0:00:00.000000"
-                    }
-                ], 
-                "timeid": "151023-114939", 
-                "loopid": "1", 
-                "total_exec_time": "0:00:29.712507"
-            }, 
-            {
-                "timeid": "151023-115008", 
-                "scenario": {
-                    "fileTinkerDownload_filesize": 300000000, 
-                    "fileTinkerDownload_tinker_wait": 4
-                }, 
-                "loopid": "1", 
-                "scenarioid": "3", 
-                "results": [
-                    {
-                        "worker": "worker0", 
-                        "sync_time": "0:00:14.828000"
-                    }, 
-                    {
-                        "worker": "worker1", 
-                        "sync_time": "0:00:12.596000"
-                    }, 
-                    {
-                        "errors": [
-                            " error_check(md5sum(fn) == md5) failed in tinkerer() [\"/home/piotr/DeIC/smashbox-reporter/lib/test_fileTinkerDownload.py\" at line 68]"
-                        ], 
-                        "worker": "tinkerer", 
-                        "sync_time": "0:00:00.000000"
-                    }
-                ], 
-                "total_exec_time": "0:00:30.780125"
-            }
-        ]
-    }, 
-    "runid": "151023-114836"
-}
